+++ src/vendorcode/amd/opensil/genoa_poc/memmap.c
@@ -87,7 +87,7 @@ BOOT_STATE_INIT_ENTRY(BS_DEV_RESOURCES, BS_ON_ENTRY, print_memory_holes, NULL);
 void add_opensil_memmap(struct device *dev, unsigned long *idx)
 {
 	// Account for UMA and TSEG
-	const uint32_t mem_usable = (uintptr_t)cbmem_top();
+	const uint32_t mem_usable = cbmem_top();
 	const uint32_t top_mem = ALIGN_DOWN(get_top_of_mem_below_4gb(), 1 * MiB);
 	if (mem_usable != top_mem)
 		reserved_ram_from_to(dev, (*idx)++, mem_usable, top_mem);
